Update index.md
General update of all the information in connection to Deliverable 1.1.

We have removed the sequence diagram that were still visible and tried to replace them with the updated sequence diagrams in the last pull request. Unfortunately, Sequence Diagrams are not visible, probably for the naming of each file. We ask to edit the names and connect the sequence diagrams.

We are going to send the updated UC Diagram within tomorrow. We ask to put UC Diagram as first Diagram because it represent the main actors.